Major changes to move render and update calls into Application.
In Application: change mkFunction to take an AppSpec (because there was no good way to create an Application), rename update to updateFn (because it kept getting shadowed), add functions to call updateFn and renderFn and to get CSS based on window style.

In Canvas: keep application instead of renderFn and viewBox, output when state updates instead of when frame starts, try to update state right before rendering, and emit tick or not depending on output of App's render function.