Change how Interactions are processed so they are no longer pre-empted by an update function.
Instead of running interactions immediately after they are triggered, they are added to a queue. This main queue builds up interactions until the next tick (rAF call). Right before calling rAF, any interactions in the main queue are moved to the processing queue, which is passed into rAF. In rAF, the interactions are all processed before the updateFn, tracking any state changes, and if an interaction or the updateFn changes the state, a StateUpdated action is emitted. Then, an InteractionsProcessed action is emitted and the processing queue is emptied.

In the case where rAF needs to fire multiple times (if delta, scaler, or context is Nothing), the InteractionsProcessed action does not fire, so the processing queue is left alone, with new queued interactions being added to it before each tick, until the update/render function runs successfully.