Time: Add function to timestamp UpdateFunctions; simplify types
There were too many aliases and phantom types, and the newtypes in the Delta record were useless since they can't be unwrapped outside this module. Change to one newtype for current and one for previous. Change FFI type signatures to use the newtypes directly instead of wrapping primitives. Change toPrev to elapse because it sounds better