Canvas: Refactor initialize function. Remove 'notify' from app state and have it passed continually between FirstTick and Tick actions
'notify' didn't really make sense in the 'dom' record, and putting it in 'subscriptions' makes FirstTick and Tick more complicated, because they would never check 'subscriptions' otherwise. In either case, or on its own in the first level of the app state, it would always be a 'Maybe', even though it's created in 'handleAction Initialize' before any other action. Passing it directly from 'initialize' to FirstTick to Tick keeps it 1. always a concrete value, not Maybe, and 2. out of records where it doesn't quite fit.